requirements.txt
flask==1.0.2
jsonschema==3.0.1
filelock==3.0.10
opencv-python-headless==4.1.0.25
python-gnupg==0.4.4
waitress==1.2.1
